(#109): mutation 훅으로 분리
