+++ src/entities/bookmark/queries.ts
@@ -2,7 +2,8 @@ import { queryOptions } from '@tanstack/react-query';
 
 import { getBookmark } from '@/entities/bookmark';
 
-import { type GetBookmarkRequest } from '@/entities/bookmark';
+import type { GetBookmarkRequest } from '@/entities/bookmark';
+
 export const bookmarkOptions = {
   getBookmark: ({ userId, contentId }: GetBookmarkRequest) =>
     queryOptions({
