+++ src/entities/bookmark/type/request.ts
@@ -2,3 +2,7 @@ export interface GetBookmarkRequest {
   userId: string;
   contentId: string;
 }
+
+export interface ToggleBookmarkRequest extends GetBookmarkRequest {
+  bookmarked: boolean;
+}
