+++ src/features/bookmark/index.ts
@@ -1,1 +1,2 @@
 export * from './ui';
+export * from './model';
