+++ src/features/bookmark/model/index.ts
@@ -1,0 +1,1 @@
+export * from './useToggleBookmarkMutation';
