+++ src/shared/config/firebaseConfig.ts
